libs/extract-and-load/pyproject.toml
[project]
name = "extract-and-load"
version = "0.1.0"
description = "Typed models and loaders for extract/load configs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.25.1",
    "azure-storage-file-datalake>=12.22.0",
    "pydantic>=2.11.0",
    "pyarrow>=23.0.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/extract_and_load"]
